requirements.txt
fastapi
python-dotenv
google-generativeai
pytesseract
pillow
PyMuPDF
PyPDF2
SQLAlchemy
SQLAlchemy-Utils
pgvector
nltk
sentence-transformers